working code, update, new folder - drivers

+++ tasks/RTC_TASK/rtc_task.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 extern uint8_t rtc_init(void);
-extern int16_t rtc_read_temp(void);
+extern int8_t rtc_read_temp(void);
 extern uint8_t rtc_read_time_data(void);
 
 void rtc_thread(void *, void *, void *);
